# React/NoBrain-Notebook/Api/GoApi/Models/CreateNodeJson.cs
namespace GoApi.Models
{
    public class CreateNodeJson
    {
        public string Name { get; set; }
        public string Notes { get; set; }
        public string NodeAbove { get; set; }
        public string User { get; set; }
    }
}